+++ demo/src/app/components/+pop-box/pop-box-section.component.ts
@@ -10,6 +10,52 @@ import { PopBoxRef } from '../../../../../src/pop-box/pop-box-ref.service';
 })
 export class DemoPopBoxSectionComponent {
 
+    apiParameters = [
+        {
+            property: 'target',
+            description: '弹出组件位置计算的参照元素，会根据这个参照元素计算位置',
+            type: 'ElementRef | HTMLElement',
+            default: 'null'
+        },
+        {
+            property: 'position',
+            description: '不需要参照元素，直接传位置, 比如右击触发弹出，',
+            type: '{ top: number, left: number}',
+            default: 'null'
+        },
+        {
+            property: 'placement',
+            description: `弹出相对于参照物的位置，比如 bottom center, 空格前的字符串为 top | bottom | left | right，空格后的字符串是对齐方式，
+            当位置为 top | bottom 时，对齐方式有 center | left | right, 当位置为 left | left 时，对齐方式有 center | top | bottom`,
+            type: 'String',
+            default: 'bottom center'
+        },
+        {
+            property: 'offset',
+            description: '弹出元素和参照物之间的位移',
+            type: 'Number',
+            default: '2'
+        },
+        {
+            property: 'outsideAutoClose',
+            description: '点击外部元素自动关闭',
+            type: 'Boolean',
+            default: 'true'
+        },
+        {
+            property: 'insideAutoClose',
+            description: '点击弹出组件内部元素自动关闭',
+            type: 'Boolean',
+            default: 'false'
+        },
+        {
+            property: 'arrow',
+            description: '是否有 arrow，暂时未实现此功能',
+            type: 'Boolean',
+            default: 'false'
+        },
+    ];
+
     public demoPlacement = 'bottom center';
 
     public config = {